Working basic landmass generation
The Voronoi library function has been correctly wrapped to produce
a graph that is usable for map generation.

A simple map generator has been implemented that only creates a
landmass and assigns three classes to each polygon: Ocean, Land or
Lake.

A test program has been added so that the generated terrain can be
inspected.